Restructuring WolfSheep model
