+++ examples/WolfSheep/Readme.md
@@ -1,20 +1,34 @@
-## Wolf-Sheep Predation Model
+# Wolf-Sheep Predation Model
 
-Another ABM standard. Wolf and sheep wander around the grid randomly and reproduce. When a wolf and a sheep end up on the same grid cell, the wolf eats the sheep. 
+## Summary
 
- Implemented here to test / demonstrate several Mesa concepts and features:
+A simple ecological model, consisting of three agent types: wolves, sheep, and grass. The wolves and the sheep wander around the grid at random. Wolves and sheep both expend energy moving around, and replenish it by eating. Sheep eat grass, and wolves eat sheep if they end up on the same grid cell.
+
+If wolves and sheep have enough energy, they reproduce, creating a new wolf or sheep (in this simplified model, only one parent is needed for reproduction). The grass on each cell regrows at a constant rate. If any wolves and sheep run out of energy, they die.
+
+The model is tests and demonstrates several Mesa concepts and features:
  - MultiGrid
- - Multiple agent types (wolves, sheep, possibly grass)
+ - Multiple agent types (wolves, sheep, grass)
+ - Agents inheriting a behavior (random movement) from an abstract parent
  - Writing a model composed of multiple files.
- - Writing a generalizable behavior (in this case, random movement) 
  - Dynamically adding and removing agents from the schedule
- - (Eventually) visualizing a layered grid, with a grass layer, a wolf layer and a sheep layer.
 
-### To run this example
+## How to Run
 
-* Launch the model
-```python
-python WolfSheepServer.py
+To run the model interactively, run ``run.py`` in this directory. e.g.
+
 ```
-* Visit your browser: http://127.0.0.1:8888/
-* In your browser hit *reset*, then *run*
+    $ python run.py
+``` 
+
+Then open your browser to [http://127.0.0.1:8888/](http://127.0.0.1:8888/) and press Reset, then Run. 
+
+## Files
+
+* ``wolf_sheep/random_walker.py``: This defines the ``RandomWalker`` agent, which implements the behavior of moving randomly across a grid, one cell at a time. Both the Wolf and Sheep agents will inherit from it.
+* ``wolf_sheep/test_random_walk.py``: Defines a simple model and a text-only visualization intended to make sure the RandomWalk class was working as expected. This doesn't actually model anything, but serves as an ad-hoc unit test. To run it, ``cd`` into the ``wolf_sheep`` directory and run ``python test_random_walk.py``.
+* ``wolf_sheep/agents.py``: Defines the Wolf, Sheep, and GrassPatch agent classes.
+* ``wolf_sheep/schedule.py``: Defines a custom variant on the RandomActivation scheduler, where all agents of one class are activated (in random order) before the next class goes -- e.g. all the wolves go, then all the sheep, then all the grass.
+* ``wolf_sheep/model.py``: Defines the Wolf-Sheep Predation model itself
+* ``wolf_sheep/server.py``: Sets up the interactive visualization server
+* ``run.py``: Launches a model visualization server.
