+++ examples/WolfSheep/run.py
@@ -1,0 +1,3 @@
+from wolf_sheep.server import server
+
+server.launch()
